spark/requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
py4j==0.10.9.7
pandas>=2.0.0
pyarrow>=14.0.0
psycopg2-binary
